+++ apps/spruce/src/pages/waterfall/styles.ts
@@ -1,0 +1,31 @@
+import { css } from "@emotion/react";
+import styled from "@emotion/styled";
+import { size } from "constants/tokens";
+
+const BUILD_VARIANT_WIDTH = 200;
+const INACTIVE_WIDTH = 80;
+
+export const gridGroupCss = css`
+  display: flex;
+  gap: ${size.s};
+  flex-grow: 1;
+  padding-left: ${size.xs};
+  padding-right: ${size.xs};
+`;
+
+export const BuildVariantTitle = styled.div`
+  flex-grow: 0;
+  flex-shrink: 0;
+  word-break: break-word;
+  width: ${BUILD_VARIANT_WIDTH}px;
+`;
+
+export const Row = styled.div`
+  display: flex;
+  gap: ${size.xs};
+  margin-bottom: ${size.s};
+`;
+
+export const InactiveVersion = styled.div`
+  width: ${INACTIVE_WIDTH}px;
+`;
